+++ tilematrix/__init__.py
@@ -1,19 +1,21 @@
 """Main package entry point."""
 
+from ._grid import GridDefinition
+from ._funcs import clip_geometry_to_srs_bounds, snap_bounds
 from ._tilepyramid import TilePyramid
 from ._tile import Tile
-from ._funcs import clip_geometry_to_srs_bounds, snap_bounds
 from ._types import Bounds, Shape, TileIndex
 
 __all__ = [
+    'Bounds',
+    'clip_geometry_to_srs_bounds',
+    'GridDefinition',
+    'Shape',
+    'snap_bounds',
     'TilePyramid',
     'Tile',
     'TileIndex',
-    'Shape',
-    'Bounds',
-    'clip_geometry_to_srs_bounds',
-    'snap_bounds'
 ]
 
 
-__version__ = "0.18"
+__version__ = "0.19"