+++ tilematrix/_conf.py
@@ -12,14 +12,18 @@ PYRAMID_PARAMS = {
         "shape": (1, 2),  # tile rows and columns at zoom level 0
         "bounds": (-180., -90., 180., 90.),  # pyramid bounds in pyramid CRS
         "is_global": True,  # if false, no antimeridian handling
-        "epsg": 4326  # EPSG code for CRS
-        },
+        "srs": {
+            "epsg": 4326  # EPSG code for CRS
+        }
+    },
     "mercator": {
         "shape": (1, 1),
         "bounds": (
             -20037508.3427892, -20037508.3427892, 20037508.3427892,
             20037508.3427892),
         "is_global": True,
-        "epsg": 3857
+        "srs": {
+            "epsg": 3857
         }
     }
+}