+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 #########
 
+0.19
+----
+* ``TilePyramid.srid`` and ``TilePyramid.type``  are deprecated
+* ``GridDefinition`` can now be loaded from package root
+* ``GridDefinition`` got ``to_dict()`` and ``from_dict()``
+
 0.18
 ----
 * order of ``Tile.shape`` swapped to ``(height, width)`` in order to match rasterio array interpretation